+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-0.1.1
+0.2.0
+
+- (feature) add undocumented params to BuildsAPI#build_list: `page`, `per_page`
+- (feature) add undocumented param to ReposAPI#repo_delete: `remove`
+- (feature) add undocumented params to ReposAPI#repo_list: `page`, `per_page`
+- (feature) add undocumented endpoint to UserAPI: `user_token`
+- (dependency) update octokit from `~> 5.5` to `~> 5.6`
+
+  0.1.1
 
 - (feature) Added documentation to API methods.
 - (fix) removed required but unused argument from ReposAPI#repo_list